Reimplement failure-injection using an ExecutionInterceptor
